added persistence

+++ app/src/main/java/com/volfor/ondori/features/alarm/domain/usecases/CancelAlarmUseCase.kt
@@ -1,12 +1,12 @@
 package com.volfor.ondori.features.alarm.domain.usecases
 
-import com.volfor.ondori.features.alarm.domain.AlarmScheduler
+import com.volfor.ondori.features.alarm.domain.scheduler.AlarmScheduler
 import javax.inject.Inject
 
 class CancelAlarmUseCase @Inject constructor(
     private val alarmScheduler: AlarmScheduler
 ) {
-    operator fun invoke(alarmId: Int) {
-        alarmScheduler.cancelAlarm(alarmId)
+    operator fun invoke(alarmId: Long) {
+        alarmScheduler.cancelAlarm(alarmId.toInt())
     }
 }
